--- client/src/components/Proverbs/proverb.css
@@ -1,20 +1,0 @@
-.input{
-width: 100%;
-height: 200px;
-border-radius: 20px;
-}
-.btns{
-   width:35%;
-    display: flex;
-    flex-direction: row;
-    justify-content:space-around;
-    align-self: center;
-
-}
-.proverbcontainer{
-    width:100%;
-    height: 270px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-}
