+++ client/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import { ConfigProvider } from 'antd';
 import App from './App/index';
-import Addproverbs from './components/Proverbs/Addproverbs';
+import Addproverbs from './components/Proverbs';
 
 ReactDOM.render(
   <ConfigProvider direction="rtl">
